+++ src/colorlightservice.ts
@@ -6,7 +6,7 @@ import {
   POWERMODE_CT,
   powerModeFromColorModeAndActiveMode,
   Attributes,
-  ConcreteLightService,
+  ConcreteLightService
 } from "./lightservice";
 
 export class ColorLightService extends LightService implements ConcreteLightService {
@@ -20,21 +20,21 @@ export class ColorLightService extends LightService implements ConcreteLightServ
     this.handleCharacteristic(
       this.platform.Characteristic.On,
       async () => this.getAttribute("power"),
-      value => this.sendCommand("set_power", [value ? "on" : "off", "smooth", 500, 0]),
+      (value) => this.sendCommand("set_power", [value ? "on" : "off", "smooth", 500, 0])
     );
     this.handleCharacteristic(
       this.platform.Characteristic.Brightness,
       async () => this.getAttribute("bright"),
-      async value => {
-        this.log("set brightness", value);
-        if (value > 0) { 
+      async (value) => {
+        this.log(`set brightness to ${value}`);
+        if (value > 0) {
           await this.sendSuddenCommand("set_bright", value);
         } else {
           await this.sendSuddenCommand("set_power", "off");
           this.log(`set brightness to 0, power off`);
         }
         this.saveDefaultIfNeeded();
-      },
+      }
     );
     if (this.platform.config.ctforcolor) {
       const characteristic = this.handleCharacteristic(
@@ -44,7 +44,7 @@ export class ColorLightService extends LightService implements ConcreteLightServ
           this.debug(`getCT: ${JSON.stringify(attributes)} -> ${convertColorTemperature(attributes.ct)}`);
           return convertColorTemperature(attributes.ct);
         },
-        async value => {
+        async (value) => {
           const kelvin = convertColorTemperature(value);
           await this.ensurePowerMode(POWERMODE_CT);
           this.debug(`setCT: ${convertColorTemperature(value)}`);
@@ -52,12 +52,12 @@ export class ColorLightService extends LightService implements ConcreteLightServ
           this.setAttributes({ ct: kelvin });
           this.updateColorFromCT(value);
           this.saveDefaultIfNeeded();
-        },
+        }
       );
       characteristic.setProps({
         ...characteristic.props,
         maxValue: convertColorTemperature(this.specs.colorTemperature.min),
-        minValue: convertColorTemperature(this.specs.colorTemperature.max),
+        minValue: convertColorTemperature(this.specs.colorTemperature.max)
       });
     }
     this.handleCharacteristic(
@@ -67,10 +67,10 @@ export class ColorLightService extends LightService implements ConcreteLightServ
         this.log(`getHue: ${JSON.stringify(attributes)}`);
         return attributes.hue;
       },
-      async value => {
+      async (value) => {
         this.lastHue = value;
         await this.setHSV();
-      },
+      }
     );
     this.handleCharacteristic(
       this.platform.Characteristic.Saturation,
@@ -79,10 +79,10 @@ export class ColorLightService extends LightService implements ConcreteLightServ
         this.debug(`getSat: ${JSON.stringify(attributes)}`);
         return attributes.sat;
       },
-      async value => {
+      async (value) => {
         this.lastSat = value;
         await this.setHSV();
-      },
+      }
     );
   }
 
@@ -97,7 +97,7 @@ export class ColorLightService extends LightService implements ConcreteLightServ
     if (this.powerMode === POWERMODE_CT) {
       this.updateCharacteristic(
         this.platform.Characteristic.ColorTemperature,
-        convertColorTemperature(newAttributes.ct),
+        convertColorTemperature(newAttributes.ct)
       );
     }
     this.updateCharacteristic(this.platform.Characteristic.On, newAttributes.power);