+++ src/index.ts
@@ -1,7 +1,7 @@
 import { API } from "homebridge";
 
 import { PLATFORM_NAME } from "./settings";
-import { YeelighterPlatform } from "./platform"; 
+import { YeelighterPlatform } from "./platform";
 
 /**
  * This method registers the platform with Homebridge